timeout v3

+++ Sender.java
@@ -12,7 +12,6 @@ import java.util.PriorityQueue;
 import java.util.concurrent.ConcurrentLinkedDeque;
 import java.util.concurrent.locks.ReentrantLock;
 
-import javax.swing.text.Segment;
 
 import java.util.Map;
 import java.util.HashMap;
@@ -61,7 +60,7 @@ public class Sender {
 
     private class SenderTimeout implements Runnable {
         
-        int localTimeout = timeout;
+        long localTimeout = timeout;
 
         @Override
         public void run() {
